refactor(app): move customer observer to a package
